Phase 6: Add block transformation events
Summary: This commit implements Phase 6 of the API parity plan, adding five new block transformation events and wiring them into the appropriate game logic paths. All events are exposed to native plugins and WASM guests and are cancellable.

New event definitions (pumpkin/src/plugin/api/events/block/ and entity/):
- BlockFormEvent: fired when a block is formed naturally; wired in lava flow (stone/cobblestone), coral death, sponge drying, and farmland moisture loss.
- BlockMultiPlaceEvent: fired when a placement creates multiple blocks; wired in BlockRegistry::get_multi_place_positions for beds, doors, tall plants, and small dripleaf.
- StructureGrowEvent: fired when a structure grows from a sapling; wired in SaplingBlock::generate.
- EntityBlockFormEvent: fired when an entity creates a block; wired in Enderman block placement, small fireball impact, and falling-block landing.
- EntityChangeBlockEvent: fired when an entity modifies a block; wired in Enderman block pickup, sheep eating grass, Ender Dragon block destruction, and splash/lingering potion extinguishing fire.

WASM / plugin-api parity:
- Added WIT definitions for the five events in pumpkin-plugin-wit/v0.1/event.wit.
- Added host conversions in pumpkin/src/plugin/loader/wasm/wasm_host/wit/v0_1/events/block.rs and entity.rs.
- Registered event routing in context.rs.
- Added plugin-api marker structs in pumpkin-plugin-api/src/events/.

Other changes:
- Registered all Phase 6 modules and re-exports in the events block/entity mod.rs files.
- Added phase6_tests.rs covering cancellation and field mutability for the new events.
- Fixed pre-existing clippy warnings surfaced by cargo clippy --all-targets --all-features and ran cargo fmt.
- All pumpkin lib tests pass (166 passed).

+++ pumpkin/src/entity/item.rs
@@ -549,7 +549,7 @@ impl EntityBase for ItemEntity {
                             .get_item_entity()
                             .expect("entity should be an item"),
                         item_stack,
-                        count_before as u8,
+                        count_before,
                     ))
                     .await;
                 if event.cancelled {
